==> Notebooks/3. EDA.ipynb <==
import pandas as pd
from wordcloud import WordCloud
import matplotlib.pyplot as plt
from textwrap import wrap

from google.colab import drive
drive.mount('/content/drive')

# astrazeneca = pd.read_csv('/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/Processed/astrazeneca.csv')
# moderna = pd.read_csv('/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/Processed/moderna.csv')
# efektif = pd.read_csv('/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/Processed/efektif.csv')
# pfizer = pd.read_csv('/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/Processed/pfizer.csv')
# sinopharm = pd.read_csv('/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/Processed/sinopharm.csv')
# sinovac = pd.read_csv('/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/Processed/sinopharm.csv')

import os
import mimetypes
import pandas as pd
input_path = "/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/2. Clean/Twitter/v3"

data = []
title = []

list_data = os.listdir(input_path)
for i in list_data:
  cur_file = input_path + f"/{i}"
  if mimetypes.guess_type(cur_file)[0] != "text/csv":
      continue
  title.append(i)
  data.append(pd.read_csv(cur_file))

# data = [astrazeneca,moderna,efektif,pfizer,sinopharm,sinovac]
# title = ['astrazeneca','moderna','efektif','pfizer','sinopharm','sinovac']

astrazeneca

# Function for generating word clouds
def generate_wordcloud(data,title):
  wc = WordCloud(width=400, height=330, max_words=100,colormap="Dark2",scale=2).generate(data.to_string())
  plt.figure(figsize=(10,8))
  plt.imshow(wc, interpolation='bilinear')
  plt.title(title)
  plt.axis("off")
  plt.show()


for i,df in enumerate(data):
  generate_wordcloud(df['raw'],title[i])
  generate_wordcloud(df['clean'],title[i])
  print('-------------------------------------')
