==> Notebooks/4. Pre and Post Labeling.ipynb <==
import pandas as pd
import os
import mimetypes

from google.colab import drive
drive.mount('/content/gdrive')

# Pre Labeling
Define Script Variables

"""
Script ini digunakan untuk menggabungkan beberapa file csv di dalam suatu folder
dan akan mengeluarkan 1 csv gabungan yang diacak.

@params
input_path: string
            contoh: "/path/menuju/folder/banyak/csv"
csv_output_path: string
                 contoh: "/path/menuju/output.csv"
"""

input_path = "/content/gdrive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/2. Clean/Twitter/v3"
csv_output_path = "/content/gdrive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/3. Labeling/shuffle_v3.csv"

def concat_and_shuffle(input_dir:str, output_file:str) -> None:
  list_file = os.listdir(input_path)
  all_df = []
  for i in list_file:
    cur_file = input_path+f"/{i}"
    if mimetypes.guess_type(cur_file)[0] != "text/csv":
      continue
    all_df.append(pd.read_csv(cur_file))

  final_df = pd.concat(all_df)
  final_df = final_df.sample(len(final_df))
  print("Total data:", len(final_df))
  print("data has been concatenated and has been shuffled and",
        f"outputted to '{csv_output_path}'")
  final_df.reset_index(drop=True).to_csv(output_file)
  
concat_and_shuffle(input_path, csv_output_path)

# Post Labeling

wait ini gw buat untuk apaan dah? - kaenova

"""
reference: https://datascience.stackexchange.com/questions/37227/how-to-remove-rows-from-a-data-frame-that-are-identical-to-other-df
Script ini digunakan untuk menghapus data yang ada pada satu csv file ke banyak csv file lainnya

@params
input_csv_path: string
                contoh: "/path/menuju/satu/file.csv"
replace_dir_path: string
                  contoh: "/path/menuju/folder/banyak/csv"
out_path: string
          contoh: "/path/menuju/folder/banyak/csv"

anchor_column: string
               context: Kolom yang akan menjadi anchor yang sama pada pada input_csv dan banyak csv pada replace_dir_path
               contoh: "same_column"
"""

input_csv_path = "/content/gdrive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/3. Labeling/shuffle_v2.csv"
replace_dir_path = "/content/gdrive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/2. Clean/v2/improved"
out_path = "/content/gdrive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/4. Train"
anchor_column = "clean"

def del_same_row_1csv_to_many_csv(input_csv:str, many_csv_dir:str,
                                  out_dir:str, anchor: str) -> None:
  if not os.path.exists(out_dir):
    os.mkdir(out_dir)
  
  anchor_df = pd.read_csv(input_csv)

  list_file = os.listdir(input_csv)
  for i in list_file:
    cur_file = input_csv+f"/{i}"
    if mimetypes.guess_type(cur_file)[0] != "text/csv":
      continue
    cur_df = pd.read_csv(cur_file)
    cond = cur_df[anchor].isin(anchor_df[anchor])
    cur_df = cur_df.drop(cur_df[cond].index)
    cur_df.to_csv(f"{out_dir}/{i}",index=False)
    print(f"{i} has been processed and outputted to {out_dir}/{i}")

del_same_row_1csv_to_many_csv(input_csv_path, replace_dir_path, )
