==> Notebooks/5. Data Train,Validation Test Split.ipynb <==
# Manual Labels
!gdown 1rqq4u_wcqW6qXtCjh6IkXODaVuGXyTkY
# General Training
!gdown 107tWzCwhkRtHGynWmSoEMNxR2NY3FnRI
!gdown 11a7m1ksllFdx1XEFH9p5xwanIYszo-tj

import pandas as pd
import numpy as np
import typing

# Preprocessing Manual Labeling
df_manual = pd.read_csv("/content/shuffle_v3_Gabungan Voting.csv").rename({"Voting" : "labels"}, axis=1)

# Preprocessing General Training Data
def preprocess_general_training(df_in: pd.DataFrame) -> pd.DataFrame:
  df = df_in.copy()
  df = df[(df["labels"] == 1) | (df["labels"] == -1)]
  df = df.reset_index(drop=True)
  df["labels"] = df["labels"].replace(-1, 0)
  df = df.sample(len(df)).reset_index(drop=True)
  return df
  

df_general_nlu = preprocess_general_training(pd.read_csv("/content/IndoNLU.csv"))
df_general_ridlife = preprocess_general_training(pd.read_csv("/content/Ridlife.csv"))

df_manual

df_general_nlu

df_general_ridlife

def print_labels(df: pd.DataFrame, name:str = None) -> None:
  labels = {}
  label_list = df["labels"].to_list()
  for i in label_list:
    if i not in labels:
      labels[i] = 0
    labels[i] += 1

  print(f"{name if name is not None else 'Labels'}:",labels)

print_labels(df_manual, "Manual Labeling")
print_labels(df_general_nlu, "IndoNLU")
print_labels(df_general_ridlife, "RiDLife")

# Dengan label seperti ini lebih baik bagaimana?
# Manual Labeling di UnderSampling untuk Testing
# Undersampling untuk IndoNLU dan RiDLife
# Buat 3 Dataset Random

from google.colab import drive
drive.mount('/content/drive')

# Undersampling Manual Labeling for Data Testing
df_manual_sampled = pd.concat([df_manual[df_manual["labels"] == 1].sample(268), df_manual[df_manual["labels"] == 0].sample(268)]).sample(268*2).reset_index(drop=True).rename(columns={"clean":"text"})
len(df_manual_sampled)

# Create Train Test

## Label Data out of Train Data

# Create data Training and Validation
from sklearn.model_selection import train_test_split
validation_size = 0.2

total_negatives = 3412 + 2687
total_positives = 6382 + 2479
print(total_negatives, total_positives)

df_train = pd.concat([df_general_nlu, df_general_ridlife])
df_train_sampled = pd.concat([df_train[df_train["labels"] == 1].sample(total_negatives), df_train[df_train["labels"] == 0]\
                              .sample(total_negatives)]).rename(columns={"clean":"text"}).sample(total_negatives*2)\
                              .reset_index(drop=True)

X_train, X_test, y_train, y_test = train_test_split(df_train_sampled[["raw", "text"]], df_train_sampled[["labels"]], test_size=validation_size)

df_train = pd.concat([X_train.reset_index(drop=True), y_train.reset_index(drop=True)], axis=1)
df_test = pd.concat([X_test.reset_index(drop=True), y_test.reset_index(drop=True)], axis=1)
df_train.to_csv("/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/4. Train, Test/train/train_4.csv", index=False)
df_test.to_csv("/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/4. Train, Test/validation/validation_4.csv", index=False)

df_train

df_train.info()

## Label on Train 50%

train_percent = 0.5

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(df_manual_sampled[["raw", "text"]], df_manual_sampled[["labels"]], stratify=df_manual_sampled["labels"], test_size=1 - train_percent)
df_manual_train = pd.concat([X_train, y_train], axis=1).reset_index(drop=True)
df_manual_test = pd.concat([X_test, y_test], axis=1).reset_index(drop=True)

# combine with train data

## Undersampling IndoNLU
sample_target_label = 3412
df_nlu_pos = df_general_nlu[df_general_nlu["labels"] == 1].sample(sample_target_label)
df_nlu_neg = df_general_nlu[df_general_nlu["labels"] == 0].sample(sample_target_label)
# Need to rename for models
df_nlu_train = pd.concat([df_nlu_pos, df_nlu_neg]).sample(sample_target_label * 2).reset_index(drop=True).rename(columns={"clean" : "text"})

## Undersampling RidLife
sample_target_label = 2479
df_ridlife_pos = df_general_ridlife[df_general_ridlife["labels"] == 1].sample(sample_target_label)
df_ridlife_neg = df_general_ridlife[df_general_ridlife["labels"] == 0].sample(sample_target_label)
# Need to rename for models
df_ridlife_train = pd.concat([df_ridlife_pos, df_ridlife_neg]).sample(sample_target_label * 2).reset_index(drop=True).rename(columns={"clean" : "text"})

# Concat IndoNLU, RidLife, and Manual Train for Training Data
pd.concat([df_nlu_train, df_ridlife_train, df_manual_train]).to_csv("/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/4.1 Train Test Baru/50 Manual/train.csv", index=False)
df_manual_test.to_csv("/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/4.1 Train Test Baru/50 Manual/test.csv", index=False)


## Label on Train 25%

train_percent = 0.25

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(df_manual_sampled[["raw", "text"]], df_manual_sampled[["labels"]], stratify=df_manual_sampled["labels"], test_size= 1-train_percent)
df_manual_train = pd.concat([X_train, y_train], axis=1).reset_index(drop=True)
df_manual_test = pd.concat([X_test, y_test], axis=1).reset_index(drop=True)

# combine with train data

## Undersampling IndoNLU
sample_target_label = 3412
df_nlu_pos = df_general_nlu[df_general_nlu["labels"] == 1].sample(sample_target_label)
df_nlu_neg = df_general_nlu[df_general_nlu["labels"] == 0].sample(sample_target_label)
# Need to rename for models
df_nlu_train = pd.concat([df_nlu_pos, df_nlu_neg]).sample(sample_target_label * 2).reset_index(drop=True).rename(columns={"clean" : "text"})

## Undersampling RidLife
sample_target_label = 2479
df_ridlife_pos = df_general_ridlife[df_general_ridlife["labels"] == 1].sample(sample_target_label)
df_ridlife_neg = df_general_ridlife[df_general_ridlife["labels"] == 0].sample(sample_target_label)
# Need to rename for models
df_ridlife_train = pd.concat([df_ridlife_pos, df_ridlife_neg]).sample(sample_target_label * 2).reset_index(drop=True).rename(columns={"clean" : "text"})

# Concat IndoNLU, RidLife, and Manual Train for Training Data
pd.concat([df_nlu_train, df_ridlife_train, df_manual_train]).to_csv("/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/4.1 Train Test Baru/25_75 Manual/train.csv", index=False)
df_manual_test.to_csv("/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/4.1 Train Test Baru/25_75 Manual/test.csv", index=False)

## Label on Train 75%

train_percent = 0.75

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(df_manual_sampled[["raw", "text"]], df_manual_sampled[["labels"]], stratify=df_manual_sampled["labels"], test_size= 1-train_percent)
df_manual_train = pd.concat([X_train, y_train], axis=1).reset_index(drop=True)
df_manual_test = pd.concat([X_test, y_test], axis=1).reset_index(drop=True)

# combine with train data

## Undersampling IndoNLU
sample_target_label = 3412
df_nlu_pos = df_general_nlu[df_general_nlu["labels"] == 1].sample(sample_target_label)
df_nlu_neg = df_general_nlu[df_general_nlu["labels"] == 0].sample(sample_target_label)
# Need to rename for models
df_nlu_train = pd.concat([df_nlu_pos, df_nlu_neg]).sample(sample_target_label * 2).reset_index(drop=True).rename(columns={"clean" : "text"})

## Undersampling RidLife
sample_target_label = 2479
df_ridlife_pos = df_general_ridlife[df_general_ridlife["labels"] == 1].sample(sample_target_label)
df_ridlife_neg = df_general_ridlife[df_general_ridlife["labels"] == 0].sample(sample_target_label)
# Need to rename for models
df_ridlife_train = pd.concat([df_ridlife_pos, df_ridlife_neg]).sample(sample_target_label * 2).reset_index(drop=True).rename(columns={"clean" : "text"})

# Concat IndoNLU, RidLife, and Manual Train for Training Data
pd.concat([df_nlu_train, df_ridlife_train, df_manual_train]).to_csv("/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/4.1 Train Test Baru/75_25 Manual/train.csv", index=False)
df_manual_test.to_csv("/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/4.1 Train Test Baru/75_25 Manual/test.csv", index=False)

## Train only on label 75 25

train_percent = 0.75

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(df_manual_sampled[["raw", "text"]], df_manual_sampled[["labels"]], stratify=df_manual_sampled["labels"], test_size= 1-train_percent)
df_manual_train = pd.concat([X_train, y_train], axis=1).reset_index(drop=True)
df_manual_test = pd.concat([X_test, y_test], axis=1).reset_index(drop=True)

df_manual_train.to_csv("/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/4.1 Train Test Baru/Manual Only/train.csv")
df_manual_test.to_csv("/content/drive/Shareddrives/Riset Sentimen Vaksin COVID: Yaudahlah/Riset/Data/4.1 Train Test Baru/Manual Only/test.csv")
